requirements.txt
# django-rest-framework==0.1.0
djangorestframework
django==3.2.9
coreapi
djongo
pymongo==3.12.1
drf_yasg==1.21.5